Fix upload public/private flag lookup (#1544)
* fix upload public or private flag lookup
* oracle 7 cleanup script error removing linux-firmware
* ubuntu 23.10 post install command on file that doesn't exist

+++ packer_templates/scripts/rhel/cleanup_yum.sh
@@ -19,8 +19,12 @@ package-cleanup --oldkernels --count=1 -y
 
 # Avoid ~200 meg firmware package we don't need
 # this cannot be done in the KS file so we do it here
-echo "Removing extra firmware packages"
-yum -y remove linux-firmware
+if test "$(uname -r)" == -- *el7uek*; then
+  echo "Skipping firmware removal for Oracle Linux"
+else
+  echo "Removing extra firmware packages"
+  yum -y remove linux-firmware
+fi
 
 echo "clean all package cache information"
 yum -y clean all  --enablerepo=\*;
